consolidated code and made SearchBar another component
