--- server/.env.example
@@ -1,3 +1,0 @@
-# Secret key
-# Replace 'secret' with the actual secret key from your bootcamp folder: State/26 server/utils
-SECRET=secret
